Narrow DC-EGM interp gather indices to int32; decouple kernel-mem dump from log_level
Two memory-profiling fixes for the DC-EGM solve at scale.

interp.py: the bracket indices `interp_on_padded_grid` and `locate_on_grid`
build from `searchsorted`+`clip` span the full continuation query mesh — the
dominant `jit_egm_step` working buffer at benchmark scale — and never exceed the
grid length (a few hundred). Under x64 they default to int64, doubling those
gather-index buffers for nothing. Casting them to int32 halves the s64 portion
of the egm_step temp with no effect on the gathered values; the solved value
function is bit-identical (101 EGM/interp tests green).

solve_brute.py: `LCM_LOG_KERNEL_MEMORY=1` now emits the per-kernel `[mem]`
compile-time lines at any solve `log_level`, including `"off"`. The env var is
the sole opt-in, so the dump logs at a level that always clears the logger's
threshold. This matters because the per-period NaN/Inf diagnostic — enabled at
every level except `"off"` — adds its own full-V transient at peak, which
masks the real kernel peak; profiling must run at `"off"`, and previously that
also silenced the `[mem]` lines.

+++ src/_lcm/egm/interp.py
@@ -63,11 +63,16 @@ def interp_on_padded_grid(
     """
     xp_filled = jnp.where(jnp.isnan(xp), jnp.inf, xp)
     n_valid = jnp.sum(~jnp.isnan(xp))
+    # The bracket indices span the full query mesh (the dominant egm_step
+    # working buffer at scale), and never exceed the grid length (a few
+    # hundred), so int32 holds them with vast headroom. Under x64 `searchsorted`
+    # would default to int64 and double these gather-index buffers; the cast
+    # halves them with no effect on the gathered values.
     upper = jnp.clip(
         jnp.searchsorted(xp_filled, x_query, side="right"),
         1,
         jnp.maximum(n_valid - 1, 1),
-    )
+    ).astype(jnp.int32)
     lower = upper - 1
     xp_lower = xp[lower]
     fp_lower = fp[lower]
@@ -127,11 +132,13 @@ def locate_on_grid(
 
     """
     n_nodes = grid.shape[0]
+    # int32 bracket indices: the grid has at most a few hundred nodes, so the
+    # x64-default int64 only doubles the index buffers for nothing.
     upper = jnp.clip(
         jnp.searchsorted(grid, x_query, side="right"),
         1,
         n_nodes - 1,
-    )
+    ).astype(jnp.int32)
     lower = upper - 1
     bracket_width = grid[upper] - grid[lower]
     safe_width = jnp.where(bracket_width == 0.0, 1.0, bracket_width)
